fix: 무대 칸을 원장에 올리고, 제목 y 게이트를 목록 상태에서 재게 한다 (#964)
CI 가 둘을 잡았고 둘 다 게이트가 옳았다.

① decisions:check — 규격(page-frame.ts)에 값을 더했으니 결정 기록이 있어야
한다. docs/DECISIONS.md 최상단에 남겼다: 실측 표(종전 1368×313 · 조립대
482×318 · 고친 뒤 590×428), 정한 것 둘, 그리고 기록된 반대와 반증 조건 —
「빈 상태를 무대로 만들면 그 화면만 다르게 느껴진다」(2026-08-09 소유자 지적이
정확히 그것이었다)와, 그 반대가 옳으면 무엇이 관측될지.

② page-frame.spec — 「세 목적지의 제목이 같은 y」가 스킬 232 로 갈렸다. 이
성질은 **목록을 그리는 상태**의 것이다(계기가 된 소유자 지적도 머리 문법에
대한 말이었다). 빈 상태의 스킬은 무대가 되므로 그 상태로는 머리 규격을 잴 수
없다. 그래서 스킬만 폴더를 연 상태로 재고, 그 스텁을 두 스펙이 함께 쓰도록
공용 헬퍼로 뽑았다. 성질을 지운 것이 아니라 각자 맞는 상태에서 재는 것이다.
