+++ docs/DECISIONS.md
@@ -42,6 +42,64 @@
 
 ---
 
+## 2026-08-12 — 열 것이 없을 때 **스킬 화면은 무대가 된다**, 그리고 무대 칸을 규격에 올린다
+
+**소집**: 열지 못했다 — 소유자가 이 세션에서 서브에이전트 호출을 쓰지 말라고 지시했다.
+규격 파일(`src/shared/ui/page-frame.ts`)에 값을 더하므로 원래는 「체계」 자리가 함께 정할
+일이다. **혼자 정한 규격은 규격이 아니라 취향**이라는 규율에 따라 값과 반증 조건을 남긴다.
+
+**트리거**: 규격 추가 — `PAGE_COLUMN_STAGE`.
+
+**계기 (관측)**: 소유자가 스크린샷을 보내며 반박했다 — *"이게 뭐니.. 공간적으로 딱봐도
+이상하지? 우측/하단 공백이 너무 심하고? 뭔가 다른 방안을 써야지?"*, 그리고 조립대 입구
+화면을 함께 보내며 *"이렇게 조립대같은 전략을 쓰던지"*.
+
+**선행 기록**: 2026-08-11 의 「페이지 틀에 폼 컬럼을 더한다」와 같은 화면이다. 그 기록의
+처방(좁은 칸 960 + 카드로 묶기)은 **부족했던 것으로 판명**됐다 — 뒤집는 것이 아니라 그
+위에 한 층을 더한다. 그 기록이 이미 적어 둔 실패("첫 처방은 화면이 반박했다")와 같은
+방식으로 두 번째 처방도 화면이 반박한 것이다.
+
+**실측** (1512×900, **잎 요소만** 잰 잉크 상자 — 컨테이너까지 세면 화면 전체가 나와
+「공백 0」이라는 거짓이 된다):
+
+| | 잉크 | 좌 / 우 | 상 / 하 |
+|---|---|---|---|
+| 스킬 빈 상태 (종전) | 1368×313 | 104 / 40 | 56 / **531** |
+| 조립대 입구 | 482×318 | 489 / 541 | 291 / 291 |
+| 스킬 빈 상태 (고친 뒤) | 590×428 | 493 / 429 | 239 / 233 |
+
+조립대는 **가운데에 세워져** 있고 스킬은 위에 붙어 벽까지 퍼져 있었다. 「횡하다」는
+공백의 양이 아니라 **그 덩어리가 화면에 묶여 있는가**의 문제다.
+
+**정한 것 둘**
+
+1. **열 것이 아직 없을 때 스킬 화면은 머리 행을 쓰지 않고 무대가 된다.** 제목이 무대의
+   제목이 되고, 목적지가 무엇인지는 좌측 레일이 이미 말한다. 목록이 있을 때의 머리
+   (제목 + 수 + 설명)는 2026-08-09 소유자 지적으로 맞춘 것이라 **그대로 둔다**.
+2. **`PAGE_COLUMN_STAGE = "mx-auto w-full max-w-[640px]"`** 를 규격에 올린다.
+   **새 값이 아니다** — 조립대 입구가 이미 쓰던 640 이 두 번째 소비처를 얻은 것이고,
+   이 저장소의 규율이 정확히 그 시점을 이름 붙일 때로 정해 뒀다. 두 화면이 같은 한 곳을
+   가리킨다. 이 칸은 **폭만** 정하고, 가운데 세우는 것은 그 칸을 담은 무대의 일이다.
+
+**게이트 둘이 바뀐다** — 성질을 지운 것이 아니라 **각자 맞는 상태에서 재게** 했다:
+
+- `skills-inventory.spec.ts` 의 「빈 상태」 시험이 잠그던 것이 「제목 바로 아래」였다.
+  그건 2026-08-11 회차에 **구현자가 내린 결론**이고 소유자가 화면을 보고 뒤집었다.
+  이제 「가운데에 세워졌는가」를 잠근다 — 픽셀이 아니라 뷰포트 비율(가로 10% ·
+  세로 12%)로. 세로는 카드 하나가 아니라 **무대 덩어리**로 잰다(카드만 재면 위에 제목이
+  있어 늘 치우친 것으로 나온다 — 첫 판이 215/118 로 그렇게 오판했다).
+- `page-frame.spec.ts` 의 「세 목적지의 제목이 같은 y」는 **목록을 그리는 상태**의
+  성질이다. 스킬만 폴더를 연 상태로 재도록 스텁을 공용 헬퍼로 뽑았다.
+
+**기록된 반대**: 「빈 상태를 무대로 만들면 그 화면이 다른 목적지와 다르게 느껴진다」 —
+2026-08-09 소유자 지적이 정확히 그것이었다. 지금은 **상태를 갈라서** 둘을 다 지킨다고
+보지만, 이 반대가 옳으면 그 갈라짐 자체가 문제로 나타난다.
+**반증 조건**: 소유자가 스킬 화면을 두고 *"왜 여기만 다르지"* 를 다시 말하면, 또는 빈
+상태에서 목록 상태로 넘어가는 순간의 배치 점프가 거슬린다고 지적하면 — 그때는 무대를
+버리고 머리 행 아래에서 균형을 잡는 쪽으로 되돌린다.
+**재검토**: 스킬 화면에 목적지가 하나 더 붙을 때(빈 상태를 무대로 쓰는 두 번째 화면이
+생기는 순간), 또는 위 반증 조건이 관측될 때.
+
 ## 2026-08-11 — 잠깐 뜨는 표면이 **자기 종류를 선언한다**, 그리고 페이지 틀에 폼 컬럼을 더한다
 
 **소집**: 열지 못했다 — 소유자가 이 세션에서 서브에이전트 호출을 쓰지 말라고 지시했다.
