+++ src/entities/docs-vault/data/manifest.json
@@ -6217,6 +6217,11 @@
         },
         {
           "depth": 2,
+          "text": "2026-08-12 — 열 것이 없을 때 **스킬 화면은 무대가 된다**, 그리고 무대 칸을 규격에 올린다",
+          "slug": "2026-08-12-열-것이-없을-때-스킬-화면은-무대가-된다-그리고-무대-칸을-규격에-올린다"
+        },
+        {
+          "depth": 2,
           "text": "2026-08-11 — 잠깐 뜨는 표면이 **자기 종류를 선언한다**, 그리고 페이지 틀에 폼 컬럼을 더한다",
           "slug": "2026-08-11-잠깐-뜨는-표면이-자기-종류를-선언한다-그리고-페이지-틀에-폼-컬럼을-더한다"
         },
@@ -8232,8 +8237,8 @@
         }
       ],
       "excerpt": "이 파일은 결정과 그때 진 반대 의견을 남긴다. docs/CHANGELOG.md 는 무엇이 언제 바뀌었나를 답하고, 이 파일은 왜 그렇게 정했고, 무엇을 걸었나를 답한다. 최신이 위. 기록은 덧붙이기만 한다 — 지난 기록을 고치지 않는다. 판단이 바뀌었으면 새 기록을 쓰고 옛 기록을 뒤집힘 으로 표시한다. 틀린 기록도 자산이다. 1. 소집 전에 읽는다. 카운슬을 열기 전(그리고 단독 PO 패스를 쓰기 전) 같은 표면 · 같은 질문에 대한 선행 결정이 있는지 먼저 본다. 있으면 새 패스는 그것을 ① 여전히 유효하다고 인용하거나 ② 명시적으로 뒤집는다. 조용히 다시 결정",
-      "wordCount": 99487,
-      "updatedAt": "2026-08-11",
+      "wordCount": 99992,
+      "updatedAt": "2026-08-12",
       "linksOut": [
         "slug"
       ]
